server/requirements.txt
aniso8601==9.0.1
astroplan==0.8
astropy==4.3.1
click==8.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.1
psycopg2==2.9.1
pyerfa==2.0.0
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.22
Werkzeug==2.0.1
